+++ src/ui/mod.rs
@@ -1,7 +1,7 @@
+pub mod app;
 pub mod assets;
 pub mod colors;
 pub mod components;
 pub mod models;
 pub mod rootview;
 pub mod screens;
-pub mod types;
