+++ src/ui/screens/about/view_model.rs
@@ -1,0 +1,10 @@
+use crate::ui::app::AppModels;
+use gpui::*;
+
+pub struct AboutViewModel;
+
+impl AboutViewModel {
+    pub fn new(_window: &mut Window, _cx: &mut Context<Self>, _models: &AppModels) -> Self {
+        Self
+    }
+}
